removed unused dto,  move classes to job package
